projects/x402-python-demo/requirements.txt
x402[fastapi,httpx,evm]
uvicorn
python-dotenv
eth-account
